+++ VietausWebAPI.Core/Domain/Entities/ManufacturingSchema/MfgProductionOrder.cs
@@ -33,8 +33,8 @@ namespace VietausWebAPI.Core.Domain.Entities.ManufacturingSchema
         public DateTime? ExpectedDate { get; set; }
         public DateTime RequiredDate { get; set; }
 
-        public int TotalQuantityRequest { get; set; }
-        public int? TotalQuantity { get; set; }
+        public decimal TotalQuantityRequest { get; set; }
+        public decimal? TotalQuantity { get; set; }
 
         public int? NumOfBatches { get; set; }
         public decimal UnitPriceAgreed { get; set; }
